Fix API test auth, schema override, and broken test isolation.
