+++ tests/api/test_api.py
@@ -1,0 +1,162 @@
+"""Tests for the API."""
+
+from datetime import date
+from unittest.mock import patch
+
+import pytest
+from ic_data_repo.permissions import deposit_action
+from invenio_access.permissions import ActionUsers
+from invenio_oauth2server.models import Token
+from invenio_oauth2server.proxies import current_oauth2server
+
+
+@pytest.fixture
+def user(UserFixture, app, db):
+    """An initialised user."""
+    u = UserFixture(
+        email="foo@bar.com",
+        password="password",
+    )
+    u.create(app, db)
+    return u
+
+
+@pytest.fixture
+def user_client(user, client):
+    """A client logged in as the user fixture."""
+    return user.login(client)
+
+
+@pytest.fixture
+def user_depositor(user, db):
+    """Give the user fixture permission to create deposits."""
+    db.session.add(ActionUsers.allow(deposit_action, user_id=user.id))
+    return user
+
+
+@pytest.fixture(autouse=True)
+def mock_manifest():
+    """Mock manifest to always return a value for theme.css."""
+    with patch("flask_webpackext.manifest.JinjaManifest.__getitem__") as mock:
+        mock.return_value = '<link rel="stylesheet" href="/static/dist/theme.css">'
+        yield mock
+
+
+@pytest.fixture
+def browser_headers(user_client):
+    """Headers for web browser requests, including CSRF token."""
+    csrf_token = user_client.get_cookie("csrftoken").value
+    if not csrf_token:
+        raise ValueError("CSRF token not found in cookies")
+    return {
+        "Content-Type": "application/json",
+        "X-CSRFToken": csrf_token,
+    }
+
+
+@pytest.fixture
+def api_headers(db, user_client, user_depositor):
+    """Headers for API requests, including API token."""
+    db.session.add(ActionUsers.allow(deposit_action, user_id=user_depositor.id))
+    scopes = [s[0] for s in current_oauth2server.scope_choices()]
+    token = Token.create_personal("test_token", user_depositor.id, scopes=scopes)
+    db.session.commit()
+    return {
+        "Authorization": f"Bearer {token.access_token}",
+        "Content-Type": "application/json",
+    }
+
+
+@pytest.fixture
+def metadata():
+    """Simple record metadata."""
+    return {
+        "title": "Test Record",
+        "description": "This is a test record.",
+        "creators": [
+            {
+                "person_or_org": {
+                    "type": "personal",
+                    "given_name": "Neo",
+                    "family_name": "Anderson",
+                },
+                "role": "the one",
+            },
+        ],
+    }
+
+
+@pytest.fixture
+def access():
+    """Simple access schema."""
+    return {
+        "record": "public",
+        "files": "public",
+    }
+
+
+def test_metadata_schema(
+    client, location, vocabularies, user_depositor, api_headers, metadata, access
+):
+    """Test that the metadata schema is enforced."""
+    metadata["resource_type"] = "fake_resource_type"
+    metadata["publisher"] = "Fake Publisher"
+    metadata["publication_date"] = "1970-01-01"
+    access["record"] = "restricted"
+    access["files"] = "restricted"
+    result = client.post(
+        "/records",
+        json={"metadata": metadata, "access": access},
+        headers=api_headers,
+    )
+    assert result.status_code == 201
+
+    # Test metadata policies are enforced.
+    assert result.json["metadata"]["title"] == "Test Record"
+    assert result.json["metadata"]["resource_type"]["id"] == "dataset"
+    assert (
+        result.json["metadata"]["creators"][0]["person_or_org"]["given_name"] == "Neo"
+    )
+    assert "role" not in result.json["metadata"]["creators"][0]
+    assert result.json["metadata"]["publisher"] == "Imperial College London"
+    assert result.json["metadata"]["publication_date"] == date.today().isoformat()
+
+    # Test access policies are enforced.
+    # 'record' should be reset to public, but 'files' should remain restricted.
+    assert result.json["access"]["record"] == "public"
+    assert result.json["access"]["files"] == "restricted"
+
+
+def test_metadata_schema_rights(
+    client, location, vocabularies, user_depositor, api_headers, metadata
+):
+    """Test that the rights schema is enforced."""
+    # Test that no license is accepted.
+    metadata["rights"] = []
+    result = client.post(
+        "/records",
+        json={"metadata": metadata},
+        headers=api_headers,
+    )
+    assert result.status_code == 201
+    assert "rights" not in result.json["metadata"]
+
+    # Test that a single license is accepted.
+    metadata["rights"] = [{"id": "cc0-1.0"}]
+    result = client.post(
+        "/records",
+        json={"metadata": metadata},
+        headers=api_headers,
+    )
+    assert result.status_code == 201
+    assert result.json["metadata"]["rights"][0]["id"] == "cc0-1.0"
+
+    # Test that multiple rights entries are not accepted.
+    metadata["rights"] = [{"id": "cc0-1.0"}, {"id": "cc-by-4.0"}]
+    result = client.post(
+        "/records",
+        json={"metadata": metadata},
+        headers=api_headers,
+    )
+    assert result.status_code == 201
+    assert result.json["errors"][0]["messages"][0].startswith("No more than")
